refactor: move out ExecutionPipeline into separate file
