+++ consensus/src/lib.rs
@@ -46,6 +46,7 @@ mod util;
 pub mod consensus_provider;
 /// Required by the telemetry service
 pub mod counters;
+mod execution_pipeline;
 /// AptosNet interface.
 pub mod network_interface;
 mod payload_manager;
