+++ consensus/src/state_computer.rs
@@ -6,6 +6,7 @@ use crate::{
     block_storage::tracing::{observe_block, BlockStage},
     counters,
     error::StateSyncError,
+    execution_pipeline::ExecutionPipeline,
     monitor,
     payload_manager::PayloadManager,
     state_replication::{StateComputer, StateComputerCommitCallBackType},
@@ -17,21 +18,19 @@ use anyhow::Result;
 use aptos_consensus_notifications::ConsensusNotificationSender;
 use aptos_consensus_types::{block::Block, common::Round, executed_block::ExecutedBlock};
 use aptos_crypto::HashValue;
-use aptos_executor_types::{
-    state_checkpoint_output::StateCheckpointOutput, BlockExecutorTrait, ExecutorError,
-    ExecutorResult, StateComputeResult,
-};
+use aptos_executor_types::{BlockExecutorTrait, ExecutorResult, StateComputeResult};
 use aptos_infallible::Mutex;
 use aptos_logger::prelude::*;
 use aptos_types::{
-    account_address::AccountAddress, block_executor::partitioner::ExecutableBlock,
-    contract_event::ContractEvent, epoch_state::EpochState, ledger_info::LedgerInfoWithSignatures,
-    transaction::Transaction,
+    account_address::AccountAddress, contract_event::ContractEvent, epoch_state::EpochState,
+    ledger_info::LedgerInfoWithSignatures, transaction::Transaction,
 };
 use fail::fail_point;
 use futures::{future::BoxFuture, SinkExt, StreamExt};
 use std::{boxed::Box, sync::Arc};
-use tokio::sync::{mpsc, oneshot, Mutex as AsyncMutex};
+use tokio::sync::Mutex as AsyncMutex;
+
+pub type StateComputeResultFut = BoxFuture<'static, ExecutorResult<StateComputeResult>>;
 
 type NotificationType = (
     Box<dyn FnOnce() + Send + Sync>,
@@ -480,150 +479,4 @@ async fn test_commit_sync_race() {
     assert_eq!(*recorded_commit.time.lock(), LogicalTime::new(1, 10));
     assert!(executor.sync_to(generate_li(2, 8)).await.is_ok());
     assert_eq!(*recorded_commit.time.lock(), LogicalTime::new(2, 8));
-}
-
-const PIPELINE_DEPTH: usize = 8;
-
-struct ExecutionPipeline {
-    block_tx: mpsc::Sender<ExecuteBlockCommand>,
-}
-
-impl ExecutionPipeline {
-    pub fn spawn(executor: Arc<dyn BlockExecutorTrait>, runtime: &tokio::runtime::Handle) -> Self {
-        let (block_tx, block_rx) = mpsc::channel(PIPELINE_DEPTH);
-        let (ledger_apply_tx, ledger_apply_rx) = mpsc::channel(PIPELINE_DEPTH);
-        runtime.spawn(Self::execute_stage(
-            block_rx,
-            ledger_apply_tx,
-            executor.clone(),
-        ));
-        runtime.spawn(Self::ledger_apply_stage(ledger_apply_rx, executor));
-        Self { block_tx }
-    }
-
-    pub async fn queue(
-        &self,
-        block: ExecutableBlock,
-        parent_block_id: HashValue,
-        maybe_block_gas_limit: Option<u64>,
-    ) -> StateComputeResultFut {
-        let (result_tx, result_rx) = oneshot::channel();
-        let block_id = block.block_id;
-        self.block_tx
-            .send(ExecuteBlockCommand {
-                block,
-                parent_block_id,
-                maybe_block_gas_limit,
-                result_tx,
-            })
-            .await
-            .expect("Failed to send block to execution pipeline.");
-
-        Box::pin(async move {
-            result_rx
-                .await
-                .map_err(|err| ExecutorError::InternalError {
-                    error: format!(
-                        "Failed to receive execution result for block {}: {:?}.",
-                        block_id, err
-                    ),
-                })?
-        })
-    }
-
-    async fn execute_stage(
-        mut block_rx: mpsc::Receiver<ExecuteBlockCommand>,
-        ledger_apply_tx: mpsc::Sender<LedgerApplyCommand>,
-        executor: Arc<dyn BlockExecutorTrait>,
-    ) {
-        while let Some(ExecuteBlockCommand {
-            block,
-            parent_block_id,
-            maybe_block_gas_limit,
-            result_tx,
-        }) = block_rx.recv().await
-        {
-            let block_id = block.block_id;
-            debug!("execute_stage received block {}.", block_id);
-            let executor = executor.clone();
-            let state_checkpoint_output = monitor!(
-                "execute_block",
-                tokio::task::spawn_blocking(move || {
-                    fail_point!("consensus::compute", |_| {
-                        Err(ExecutorError::InternalError {
-                            error: "Injected error in compute".into(),
-                        })
-                    });
-                    executor.execute_and_state_checkpoint(
-                        block,
-                        parent_block_id,
-                        maybe_block_gas_limit,
-                    )
-                })
-                .await
-            )
-            .expect("Failed to spawn_blocking.");
-
-            ledger_apply_tx
-                .send(LedgerApplyCommand {
-                    block_id,
-                    parent_block_id,
-                    state_checkpoint_output,
-                    result_tx,
-                })
-                .await
-                .expect("Failed to send block to ledger_apply stage.");
-        }
-        debug!("execute_stage quitting.");
-    }
-
-    async fn ledger_apply_stage(
-        mut block_rx: mpsc::Receiver<LedgerApplyCommand>,
-        executor: Arc<dyn BlockExecutorTrait>,
-    ) {
-        while let Some(LedgerApplyCommand {
-            block_id,
-            parent_block_id,
-            state_checkpoint_output,
-            result_tx,
-        }) = block_rx.recv().await
-        {
-            debug!("ledger_apply stage received block {}.", block_id);
-            let res = async {
-                let executor = executor.clone();
-                monitor!(
-                    "ledger_apply",
-                    tokio::task::spawn_blocking(move || {
-                        executor.ledger_update(block_id, parent_block_id, state_checkpoint_output?)
-                    })
-                )
-                .await
-                .expect("Failed to spawn_blocking().")
-            }
-            .await;
-            result_tx.send(res).unwrap_or_else(|err| {
-                error!(
-                    block_id = block_id,
-                    "Failed to send back execution result for block {}: {:?}", block_id, err,
-                );
-            });
-        }
-        debug!("ledger_apply stage quitting.");
-    }
-}
-
-pub type StateComputeResultFut = BoxFuture<'static, ExecutorResult<StateComputeResult>>;
-
-struct ExecuteBlockCommand {
-    block: ExecutableBlock,
-    parent_block_id: HashValue,
-    maybe_block_gas_limit: Option<u64>,
-    result_tx: oneshot::Sender<ExecutorResult<StateComputeResult>>,
-}
-
-struct LedgerApplyCommand {
-    block_id: HashValue,
-    parent_block_id: HashValue,
-    state_checkpoint_output: ExecutorResult<StateCheckpointOutput>,
-    result_tx: oneshot::Sender<ExecutorResult<StateComputeResult>>,
 }
